traces/stf_trace_archive/src/requirements.txt
pandas==2.3.1
PyYAML
